// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
      vessels: {
        Row: {
          id: string;
          company_id: string | null;
          user_id: string | null;
          name: string;
          registration_number: string;
          type: string;
          captain_name: string | null;
          captain_user_id: string | null;
          capacity: number | null;
          length: number | null;
          created_at: string;
          crew_count: number | null;
          fishing_method: string | null;
          fishing_gear: Record<string, unknown> | null;
          engine_power: string | null;
          owner_name: string | null;
          width: string | null;
          fishery_permit: string | null;
          expiration_date: string | null;
          fileUrl: string | null;
          materials: string | null;
          draught: string | null;
          gross_tonnage: string | null;
          number_engines: string | null;
          type_of_machine: string | null;
          port_registry: string | null;
          owner_id: string | null;
          type_of_vessel: string | null;
          residential_address: string | null;
          image_url: string | null;
          latitude: string | null;
          longitude: string | null;
          current_zone: string | null;
        };
        Insert: {
          id?: string;
          company_id?: string | null;
          user_id?: string | null;
          name: string;
          registration_number: string;
          type: string;
          captain_name?: string | null;
          captain_user_id?: string | null;
          capacity?: number | null;
          length?: number | null;
          created_at?: string;
          crew_count?: number | null;
          fishing_method?: string | null;
          fishing_gear?: Record<string, unknown> | null;
          engine_power?: string | null;
          owner_name?: string | null;
          width?: string | null;
          fishery_permit?: string | null;
          expiration_date?: string | null;
          fileUrl?: string | null;
          materials?: string | null;
          draught?: string | null;
          gross_tonnage?: string | null;
          number_engines?: string | null;
          type_of_machine?: string | null;
          port_registry?: string | null;
          owner_id?: string | null;
          type_of_vessel?: string | null;
          residential_address?: string | null;
          image_url?: string | null;
          latitude?: string | null;
          longitude?: string | null;
          current_zone?: string | null;
        };
        Update: {
          id?: string;
          company_id?: string | null;
          user_id?: string | null;
          name?: string;
          registration_number?: string;
          type?: string;
          captain_name?: string | null;
          captain_user_id?: string | null;
          capacity?: number | null;
          length?: number | null;
          created_at?: string;
          crew_count?: number | null;
          fishing_method?: string | null;
          fishing_gear?: Record<string, unknown> | null;
          engine_power?: string | null;
          owner_name?: string | null;
          width?: string | null;
          fishery_permit?: string | null;
          expiration_date?: string | null;
          fileUrl?: string | null;
          materials?: string | null;
          draught?: string | null;
          gross_tonnage?: string | null;
          number_engines?: string | null;
          type_of_machine?: string | null;
          port_registry?: string | null;
          owner_id?: string | null;
          type_of_vessel?: string | null;
          residential_address?: string | null;
          image_url?: string | null;
          latitude?: string | null;
          longitude?: string | null;
          current_zone?: string | null;
        };
      };
      fishing_trips: {
        Row: {
          id: string;
          trip_code: string;
          user_id: string;
          vessel_id: string | null;
          vessel: string;
          address: string | null;
          departure_date: string | null;
          return_date: string | null;
          departure_port: string | null;
          departure_port_name: string | null;
          return_port: string | null;
          status: string;
          form_code: string;
          created_at: string;
          departure_province: string | null;
          place_of_departure: string | null;
          to_region: string | null;
          trip_period: string | null;
          number_of_crew: string | null;
          vessel_type: string | null;
          vessel_registration_number: string | null;
          docking_id: string | null;
          dock_province: string | null;
          place_of_dock: string | null;
          docking_date: string | null;
          total_trip_period: string | null;
          dock_code: string | null;
          fishing_logbook: string | null;
          trading_logbook: string | null;
          transshipment_logbook: string | null;
          owner_name: string | null;
        };
        Insert: {
          id?: string;
          vessel_id?: string | null;
          user_id: string;
          vessel: string;
          trip_code: string;
          address: string | null;
          departure_port_name: string | null;
          departure_date?: string | null;
          return_date?: string | null;
          departure_port?: string | null;
          return_port?: string | null;
          status: string;
          form_code: string;
          created_at?: string;
          departure_province?: string | null;
          place_of_departure?: string | null;
          to_region?: string | null;
          trip_period?: string | null;
          number_of_crew?: string | null;
          vessel_type?: string | null;
          vessel_registration_number?: string | null;
          docking_id?: string | null;
          dock_province?: string | null;
          place_of_dock?: string | null;
          docking_date?: string | null;
          total_trip_period?: string | null;
          dock_code?: string | null;
          fishing_logbook?: string | null;
          trading_logbook?: string | null;
          transshipment_logbook?: string | null;
          owner_name?: string | null;
        };
        Update: {
          id?: string;
          vessel_id?: string;
          user_id?: string;
          trip_code?: string;
          vessel: string;
          departure_port_name: string | null;
          address: string | null;
          departure_date?: string | null;
          return_date?: string | null;
          departure_port?: string | null;
          return_port?: string | null;
          status?: string;
          form_code?: string;
          created_at?: string;
          departure_province?: string | null;
          place_of_departure?: string | null;
          to_region?: string | null;
          trip_period?: string | null;
          number_of_crew?: string | null;
          vessel_type?: string | null;
          vessel_registration_number?: string | null;
          docking_id?: string | null;
          dock_province?: string | null;
          place_of_dock?: string | null;
          docking_date?: string | null;
          total_trip_period?: string | null;
          dock_code?: string | null;
          fishing_logbook?: string | null;
          trading_logbook?: string | null;
          transshipment_logbook?: string | null;
          owner_name?: string | null;
        };
      };
// ...
      seaports: {
        Row: {
          id: string;
          name: string;
          address: string | null;
          classification: number | null;
          province: string | null;
          district: string | null;
          ward: string | null;
          status: string | null;
          latitude: number | null;
          longitude: number | null;
          image_url: string | null;
          created_at: string | null;
          updated_at: string | null;
          number: number;
        };
        Insert: {
          id?: string;
          name: string;
          address?: string | null;
          classification?: number | null;
          province?: string | null;
          district?: string | null;
          ward?: string | null;
          status?: string | null;
          latitude?: number | null;
          longitude?: number | null;
          image_url?: string | null;
          created_at?: string | null;
          updated_at?: string | null;
          number?: number;
        };
        Update: {
          id?: string;
          name?: string;
          address?: string | null;
          classification?: number | null;
          province?: string | null;
          district?: string | null;
          ward?: string | null;
          status?: string | null;
          latitude?: number | null;
          longitude?: number | null;
          image_url?: string | null;
          created_at?: string | null;
          updated_at?: string | null;
          number?: number;
        };
      };
// ...
    };
    Views: Record<string, never>;
    Functions: Record<string, never>;
    Enums: Record<string, never>;
  };
};